Play, playnext, and playnow share common function Consolidated to a module named queueHandler.js, where processing behavior is shared instead of copied between command files.
